Move interfaces to separate folder
