+++ src/commands/interfaces/ICommand.ts
