+++ src/commands/interfaces/ICommandData.ts
